pyproject.toml
[tool.poetry]
name = "scrapybara-demos"
version = "0.1.0"
description = ""
authors = ["Justin Sun <justinsunyt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scrapybara = "2.0.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
